+++ template-service/service/src/main/java/io/harness/template/services/NGTemplateServiceImpl.java
@@ -8,7 +8,6 @@
 package io.harness.template.services;
 
 import static io.harness.annotations.dev.HarnessTeam.CDC;
-import static io.harness.beans.FeatureName.NG_SETTINGS;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 import static io.harness.exception.WingsException.USER;
@@ -31,11 +30,13 @@ import io.harness.accesscontrol.acl.api.ResourceScope;
 import io.harness.accesscontrol.clients.AccessControlClient;
 import io.harness.account.AccountClient;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.beans.FeatureName;
 import io.harness.beans.IdentifierRef;
 import io.harness.data.structure.EmptyPredicate;
 import io.harness.encryption.Scope;
 import io.harness.enforcement.client.services.EnforcementClientService;
 import io.harness.enforcement.constants.FeatureRestrictionName;
+import io.harness.engine.GovernanceService;
 import io.harness.entitysetupusageclient.remote.EntitySetupUsageClient;
 import io.harness.eraro.ErrorMessageConstants;
 import io.harness.eventsframework.schemas.entity.EntityDetailProtoDTO;
@@ -65,6 +66,7 @@ import io.harness.gitsync.persistance.GitSyncSdkService;
 import io.harness.gitsync.scm.EntityObjectIdUtils;
 import io.harness.gitsync.scm.beans.ScmCreateFileGitResponse;
 import io.harness.gitx.GitXSettingsHelper;
+import io.harness.governance.GovernanceMetadata;
 import io.harness.grpc.utils.StringValueUtils;
 import io.harness.ng.beans.PageResponse;
 import io.harness.ng.core.entitysetupusage.dto.EntitySetupUsageDTO;
@@ -75,14 +77,16 @@ import io.harness.ng.core.template.TemplateResponseDTO;
 import io.harness.ng.core.template.TemplateWithInputsResponseDTO;
 import io.harness.ngsettings.SettingIdentifiers;
 import io.harness.ngsettings.client.remote.NGSettingsClient;
+import io.harness.opaclient.model.OpaConstants;
 import io.harness.organization.remote.OrganizationClient;
 import io.harness.pms.yaml.YAMLFieldNameConstants;
 import io.harness.pms.yaml.YamlField;
+import io.harness.pms.yaml.YamlUtils;
 import io.harness.project.remote.ProjectClient;
-import io.harness.remote.client.CGRestUtils;
 import io.harness.remote.client.NGRestUtils;
 import io.harness.repositories.NGTemplateRepository;
 import io.harness.springdata.TransactionHelper;
+import io.harness.template.async.beans.SetupUsageParams;
 import io.harness.template.entity.TemplateEntity;
 import io.harness.template.entity.TemplateEntity.TemplateEntityKeys;
 import io.harness.template.events.TemplateUpdateEventType;
@@ -103,6 +107,7 @@ import io.harness.template.utils.TemplateUtils;
 import io.harness.template.yaml.TemplateRefHelper;
 import io.harness.utils.FullyQualifiedIdentifierHelper;
 import io.harness.utils.PageUtils;
+import io.harness.utils.PmsFeatureFlagService;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.inject.Inject;
@@ -149,11 +154,17 @@ public class NGTemplateServiceImpl implements NGTemplateService {
   @Inject private TemplateMergeServiceHelper templateMergeServiceHelper;
 
   @Inject private TemplateGitXService templateGitXService;
+
+  @Inject private TemplateAsyncSetupUsageService templateAsyncSetupUsageService;
+
   @Inject private GitAwareEntityHelper gitAwareEntityHelper;
   @Inject private AccountClient accountClient;
   @Inject NGSettingsClient settingsClient;
   @Inject GitXSettingsHelper gitXSettingsHelper;
   @Inject private TemplateRbacHelper templateRbacHelper;
+
+  @Inject private GovernanceService governanceService;
+  @Inject private PmsFeatureFlagService pmsFeatureFlagService;
 
   private static final String DUP_KEY_EXP_FORMAT_STRING =
       "Template [%s] of versionLabel [%s] under Project[%s], Organization [%s] already exists";
@@ -220,8 +231,12 @@ public class NGTemplateServiceImpl implements NGTemplateService {
       }
     }
 
+    checkForChildTypesInTemplates(templateEntity, "create");
+
     // apply templates to template yaml for validation and populating module info
     applyTemplatesToYamlAndValidateSchema(templateEntity);
+
+    List<EntityDetailProtoDTO> referredEntities = templateReferenceHelper.calculateTemplateReferences(templateEntity);
 
     try {
       // Check if this is template identifier first entry, for marking it as stable template.
@@ -268,7 +283,8 @@ public class NGTemplateServiceImpl implements NGTemplateService {
 
       GitAwareContextHelper.setIsDefaultBranchInGitEntityInfo();
       if (doPublishSetupUsages(template)) {
-        templateReferenceHelper.populateTemplateReferences(template);
+        templateReferenceHelper.publishTemplateReferences(
+            SetupUsageParams.builder().templateEntity(templateEntity).build(), referredEntities);
       }
 
       return template;
@@ -321,9 +337,9 @@ public class NGTemplateServiceImpl implements NGTemplateService {
 
   private void applyTemplatesToYamlAndValidateSchema(TemplateEntity templateEntity) {
     TemplateMergeResponseDTO templateMergeResponseDTO = null;
-    templateMergeResponseDTO =
-        templateMergeService.applyTemplatesToYamlV2(templateEntity.getAccountId(), templateEntity.getOrgIdentifier(),
-            templateEntity.getProjectIdentifier(), templateEntity.getYaml(), false, false, false);
+    templateMergeResponseDTO = templateMergeService.applyTemplatesToYamlV2(templateEntity.getAccountId(),
+        templateEntity.getOrgIdentifier(), templateEntity.getProjectIdentifier(),
+        YamlUtils.readAsJsonNode(templateEntity.getYaml()), false, false, false);
     populateLinkedTemplatesModules(templateEntity, templateMergeResponseDTO);
     checkLinkedTemplateAccess(templateEntity.getAccountId(), templateEntity.getOrgIdentifier(),
         templateEntity.getProjectIdentifier(), templateMergeResponseDTO);
@@ -356,6 +372,8 @@ public class NGTemplateServiceImpl implements NGTemplateService {
         templateEntity.getProjectIdentifier(), templateEntity.getRepo(), templateEntity.getConnectorRef());
     // apply templates to template yaml for validations and populating module info
     applyTemplatesToYamlAndValidateSchema(templateEntity);
+    // calculate the references, returns error if any errors occur while fetching references
+    List<EntityDetailProtoDTO> referredEntities = templateReferenceHelper.calculateTemplateReferences(templateEntity);
 
     TemplateEntity template = null;
 
@@ -368,7 +386,8 @@ public class NGTemplateServiceImpl implements NGTemplateService {
 
     GitAwareContextHelper.setIsDefaultBranchInGitEntityInfo();
     if (doPublishSetupUsages(template)) {
-      templateReferenceHelper.populateTemplateReferences(template);
+      templateReferenceHelper.publishTemplateReferences(
+          SetupUsageParams.builder().templateEntity(templateEntity).build(), referredEntities);
     }
 
     return template;
@@ -381,6 +400,8 @@ public class NGTemplateServiceImpl implements NGTemplateService {
         FeatureRestrictionName.TEMPLATE_SERVICE, templateEntity.getAccountIdentifier());
     // apply templates to template yaml for validations and populating module info
     applyTemplatesToYamlAndValidateSchema(templateEntity);
+    // calculate the references, returns error if any errors occur while fetching references
+    List<EntityDetailProtoDTO> referredEntities = templateReferenceHelper.calculateTemplateReferences(templateEntity);
 
     GitEntityInfo gitEntityInfo = GitAwareContextHelper.getGitRequestParamsInfo();
     if (gitEntityInfo != null) {
@@ -411,7 +432,8 @@ public class NGTemplateServiceImpl implements NGTemplateService {
 
     GitAwareContextHelper.setIsDefaultBranchInGitEntityInfo();
     if (doPublishSetupUsages(template)) {
-      templateReferenceHelper.populateTemplateReferences(template);
+      templateReferenceHelper.publishTemplateReferences(
+          SetupUsageParams.builder().templateEntity(templateEntity).build(), referredEntities);
     }
 
     return template;
@@ -588,7 +610,7 @@ public class NGTemplateServiceImpl implements NGTemplateService {
   }
 
   private boolean isForceDeleteEnabled(String accountIdentifier) {
-    return isNgSettingsFFEnabled(accountIdentifier) && isForceDeleteFFEnabledViaSettings(accountIdentifier);
+    return isForceDeleteFFEnabledViaSettings(accountIdentifier);
   }
 
   @VisibleForTesting
@@ -597,11 +619,6 @@ public class NGTemplateServiceImpl implements NGTemplateService {
                             .getResponse(settingsClient.getSetting(
                                 SettingIdentifiers.ENABLE_FORCE_DELETE, accountIdentifier, null, null))
                             .getValue());
-  }
-
-  @VisibleForTesting
-  protected boolean isNgSettingsFFEnabled(String accountIdentifier) {
-    return CGRestUtils.getResponse(accountClient.isFeatureFlagEnabled(NG_SETTINGS.name(), accountIdentifier));
   }
 
   private String getMessageHelper(String accountId, String orgIdentifier, String projectIdentifier) {
@@ -620,23 +637,43 @@ public class NGTemplateServiceImpl implements NGTemplateService {
       List<TemplateEntity> templateToDeleteList, Long version, String comments, boolean canDeleteStableTemplate,
       TemplateEntity stableTemplate, boolean forceDelete) {
     boolean lastUpdatedTemplateDeleted = false;
+    boolean foundStableTemplate = false;
     for (TemplateEntity templateEntity : templateToDeleteList) {
       try (TemplateGitSyncBranchContextGuard ignored = templateServiceHelper.getTemplateGitContextForGivenTemplate(
                templateEntity, GitContextHelper.getGitEntityInfo(),
                format("Deleting template with identifier [%s] and versionLabel [%s].", templateEntity.getIdentifier(),
                    templateEntity.getVersionLabel()))) {
-        if (templateEntity.isLastUpdatedTemplate()) {
-          lastUpdatedTemplateDeleted = true;
+        // If it is stable template then we will delete it at the last
+        if (templateEntity.isStableTemplate()) {
+          foundStableTemplate = true;
+          continue;
         }
         deleteSingleTemplateHelper(accountId, orgIdentifier, projectIdentifier, templateEntity.getIdentifier(),
             templateEntity, version, canDeleteStableTemplate, comments, forceDelete);
-      } catch (ReferencedEntityException referencedEntityException) {
-        if (!canDeleteStableTemplate && lastUpdatedTemplateDeleted) {
+        if (templateEntity.isLastUpdatedTemplate()) {
+          lastUpdatedTemplateDeleted = true;
+        }
+      } catch (Exception exception) {
+        // if template to delete contains stable template along with all other versions with one template having
+        // references, therefore removed !canDeleteStableTemplate from if condition.
+        if (lastUpdatedTemplateDeleted) {
           makeGivenTemplateLastUpdatedTemplateTrue(stableTemplate);
         }
-        throw referencedEntityException;
+        throw exception;
       }
     }
+
+    // Update stable Template as last updated template if the earlier lastUpdatedTemplate is Deleted.
+    if (lastUpdatedTemplateDeleted) {
+      makeGivenTemplateLastUpdatedTemplateTrue(stableTemplate);
+    }
+
+    // Delete Stable Template
+    if (canDeleteStableTemplate && foundStableTemplate) {
+      deleteSingleTemplateHelper(accountId, orgIdentifier, projectIdentifier, stableTemplate.getIdentifier(),
+          stableTemplate, version, canDeleteStableTemplate, comments, forceDelete);
+    }
+
     return true;
   }
 
@@ -822,7 +859,9 @@ public class NGTemplateServiceImpl implements NGTemplateService {
               StringValueUtils.getStringFromStringValue(templateRef.getIdentifier()),
               StringValueUtils.getStringFromStringValue(templateRef.getVersionLabel()));
 
-      unSyncedTemplate.ifPresent(templateEntity -> templateReferenceHelper.populateTemplateReferences(templateEntity));
+      unSyncedTemplate.ifPresent(templateEntity
+          -> templateReferenceHelper.populateTemplateReferences(
+              SetupUsageParams.builder().templateEntity(templateEntity).build()));
       return templateServiceHelper.makeTemplateUpdateCall(unSyncedTemplate.get(), unSyncedTemplate.get(),
           ChangeType.ADD, "", TemplateUpdateEventType.OTHERS_EVENT, true);
     } catch (DuplicateKeyException ex) {
@@ -952,6 +991,8 @@ public class NGTemplateServiceImpl implements NGTemplateService {
     TemplateEntity templateEntity =
         NGTemplateDtoMapper.toTemplateEntity(accountIdentifier, orgIdentifier, projectIdentifier, importedTemplateYAML);
 
+    checkForChildTypesInTemplates(templateEntity, "import");
+
     TemplateEntity templateEntityToSave = prepareTemplateEntity(templateEntity, repoUrl);
 
     try {
@@ -971,6 +1012,25 @@ public class NGTemplateServiceImpl implements NGTemplateService {
           format(DUP_KEY_EXP_FORMAT_STRING, templateEntity.getIdentifier(), templateImportRequest.getTemplateVersion(),
               templateEntity.getProjectIdentifier(), templateEntity.getOrgIdentifier()),
           USER_SRE, ex);
+    }
+  }
+
+  private void checkForChildTypesInTemplates(TemplateEntity templateEntity, String action) {
+    Set<TemplateEntityType> templatesWithChildTypes = new HashSet<>();
+    templatesWithChildTypes.add(TemplateEntityType.STAGE_TEMPLATE);
+    templatesWithChildTypes.add(TemplateEntityType.STEP_TEMPLATE);
+    templatesWithChildTypes.add(TemplateEntityType.STEPGROUP_TEMPLATE);
+    String error = "";
+    String actionType = action.equals("create") ? "save" : "import";
+    if (templatesWithChildTypes.contains(templateEntity.getTemplateEntityType())
+        && EmptyPredicate.isEmpty(templateEntity.getChildType())) {
+      if (templateEntity.getTemplateEntityType() == TemplateEntityType.STEPGROUP_TEMPLATE) {
+        error = "Unable to " + actionType + " the template. Missing property [stageType].";
+      } else {
+        error = "Unable to " + actionType + " the template. Missing property [type] for "
+            + templateEntity.getTemplateEntityType().toString() + " template";
+      }
+      throw new InvalidRequestException(error);
     }
   }
 
@@ -1297,7 +1357,7 @@ public class NGTemplateServiceImpl implements NGTemplateService {
     }
   }
 
-  private TemplateEntity saveTemplate(TemplateEntity templateEntity, String comments) throws InvalidRequestException {
+  TemplateEntity saveTemplate(TemplateEntity templateEntity, String comments) throws InvalidRequestException {
     if (templateServiceHelper.isOldGitSync(templateEntity)) {
       return templateRepository.saveForOldGitSync(templateEntity, comments);
     } else {
@@ -1326,6 +1386,12 @@ public class NGTemplateServiceImpl implements NGTemplateService {
           templateEntity.getIdentifier(), templateEntity.getVersionLabel(), templateEntity.getProjectIdentifier(),
           templateEntity.getOrgIdentifier(), oldTemplateEntity.getTemplateEntityType()));
     }
+
+    if (EmptyPredicate.isEmpty(oldTemplateEntity.getChildType())
+        && EmptyPredicate.isNotEmpty(templateEntity.getChildType())) {
+      return oldTemplateEntity.withChildType(templateEntity.getChildType());
+    }
+
     if (!((oldTemplateEntity.getChildType() == null && templateEntity.getChildType() == null)
             || oldTemplateEntity.getChildType().equals(templateEntity.getChildType()))) {
       throw new InvalidRequestException(format(
@@ -1409,6 +1475,12 @@ public class NGTemplateServiceImpl implements NGTemplateService {
         get(accountIdentifier, orgIdentifier, projectIdentifier, templateIdentifier, versionLabel, false, false);
 
     if (templateEntityOptional.isPresent()) {
+      if (!templateEntityOptional.get().getTemplateEntityType().isGitEntity()) {
+        throw new InvalidRequestException(String.format(
+            "Template with the given Identifier: %s and versionLabel %s cannot be moved to Git as it is not a Git Supported Template Type",
+            templateIdentifier, versionLabel));
+      }
+
       TemplateEntity movedTemplateEntity = moveTemplateEntity(accountIdentifier, orgIdentifier, projectIdentifier,
           templateIdentifier, versionLabel, moveConfigOperationDTO, templateEntityOptional.get());
 
@@ -1449,6 +1521,14 @@ public class NGTemplateServiceImpl implements NGTemplateService {
       throw new EntityNotFoundException(
           String.format("Template not found for template identifier [%s] and version label [%s] in %s",
               templateIdentifier, versionLabel, scope));
+    }
+  }
+
+  @Override
+  public void populateSetupUsageAsync(TemplateEntity templateEntity) {
+    if (templateEntity.getStoreType() == StoreType.REMOTE) {
+      SetupUsageParams setupUsageParams = SetupUsageParams.builder().templateEntity(templateEntity).build();
+      templateAsyncSetupUsageService.populateAsyncSetupUsage(setupUsageParams);
     }
   }
 
@@ -1579,5 +1659,19 @@ public class NGTemplateServiceImpl implements NGTemplateService {
     gitXSettingsHelper.setConnectorRefForRemoteEntity(accountIdentifier, orgIdentifier, projIdentifier);
     gitXSettingsHelper.setDefaultStoreTypeForEntities(
         accountIdentifier, orgIdentifier, projIdentifier, EntityType.TEMPLATE);
+  }
+
+  @Override
+  public GovernanceMetadata validateGovernanceRules(TemplateEntity templateEntity) {
+    if (!pmsFeatureFlagService.isEnabled(templateEntity.getAccountId(), FeatureName.CDS_OPA_TEMPLATE_GOVERNANCE)) {
+      return GovernanceMetadata.newBuilder()
+          .setDeny(false)
+          .setMessage(String.format("FF: [%s] is disabled for account: [%s]", FeatureName.CDS_OPA_TEMPLATE_GOVERNANCE,
+              templateEntity.getAccountId()))
+          .build();
+    }
+    return governanceService.evaluateGovernancePoliciesForTemplate(templateEntity.getYaml(),
+        templateEntity.getAccountId(), templateEntity.getOrgIdentifier(), templateEntity.getProjectIdentifier(),
+        OpaConstants.OPA_EVALUATION_ACTION_SAVE, OpaConstants.OPA_EVALUATION_TYPE_TEMPLATE);
   }
 }
